math: create handle_variable() instead of _get_implied_value()
The abstraction layers are wrong.  _get_implied_variable() is should only
handle variables instead of trying to duplicate parts of _get_value().
handle_variable() should handle const variables instead of doing that in
_get_value().  The other thing is that _get_implied_variable() calls
get_value() and it shouldn't do that.

Another thing to note is that this slightly changes how FUZZY_MIN is
handled.  The min values were always sort of bogus so I don't think this
is terrible.  I don't think this change will make a difference to the
scripts.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>