math: create hackish temporary handle_binop_rl()
I have done this in the most mindless way possible just so that I can push
the rl work down one level.  It should work exactly the same as the
original code and all the range_lists it returns will be one variable.

I'll fix it bit by bit in the following patches.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>