math: introduce new handle_subtract_rl()
The original handle_subtract() was pretty rubbish.  I can't swear that this
new totally re-written one is better, but at least it's cleaner and a step
in the right direction.

I changed a validation test.  The original version said used "s32max - 1"
and the new version says s32max.  Because really "s32max - 1" is not a
useful distinction in this case.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>