math: fix new handle_conditional_rl() code
In the old code, if it wasn't implied that the condition was true or false
then we abandoned ship and returned bogus.  My first naive attempt at
merging the true and false sides together didn't work.

But it should work now.  This is an example of how it is easier to work
with range lists than with min/max values.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>