db: fix crashing bug as a result of math changes
This is something I have been meaning to do for a while and now it causes
a crashing bug so I have to.

It's ugly to pass a "fake current slist" that should be handled
transparently in the background.  And actually it already is.  The real
current fake slist matches the slist which we pass to the call backs.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>