math: put an escape hatch in handle_conditional_rl()
In the kernel in mptbase_raid_process_event_data() there is a place which
has 10 nested conditionals.

To parse conditionals we need to break apart the condition and it's one of
the more expensive things in Smatch.  This is because you have implied
states which need to be copied over and over.

So I have put an escape hatch to give up on those cases if the memory is
low.  This is how all these cases where handled until yesterday so giving
up here is not horrible.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>