math: remove duplicate buggy code in get_implied_rl()
get_implied_rl() was horrible layering violation and because smath_math
was working with svals.  That's fixed now.

One thing to note is that I had to add a clone_rl() around the estate_rl()
when we're handling SMATCH_EXTRA states.  The reason for this is because
these range_lists are used to set other estate range_lists.  You can't have
two estate range_lists pointing to the same data or it leads to a double
free.  That code was there in the original get_implied_rl() as well.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>