comparison, db: use comparison information stored in the database
Currently, we don't load any information in the database, but I would like
to.  So if you store the return value as: "0-<1>" then it says that the
return is between zero and up to parameter one.  Currently it does this,
but it doesn't store the relationship in the comparison module so this
change adds that.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>