extra: revert part of the recurse patch which broke smatch_equiv
I removed the is_simple_math() check when I introduced has_variable()
because it seemed like it was cut and pasted from earlier.  But actually it
turns out to be needed and basically correct.

Without the check then smatch_extra was duplicating part of what
smatch_equiv does and this created bugs.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>