db, comparison, ranges: improve how comparisons will be stored in the db
I started to add this but didn't get it finished.  It's the code to get
relationship information from the database.  The code to store it in the
database has not been added yet.

The new format for storing relationships is something like:

	[==p0]

This means that the return value is equal to parameter zero.  For example
ERR_PTR() just casts parameter zero and returns it.  But a more
interesting case might look like:

	0-[<=p1]

This means that the return value is range 0 up to parameter one.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>