db, comparison: store return range comparisons in the database
This stores the relationship information in the database.  It doesn't work
very well yet, in terms of recognizing relationships in the kernel source.

What it does is that at the beginning of the function it stores all the
original values.  Then at the end of the function it compares the return
value to each parameter until it finds a relationship.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>