states: introduce __set_fake_cur_slist_fast()
This is like __set_fake_cur_slist() but it doesn't do any cloning and it's
trickier to use.  The rule is that you can't write to the cur_slist when
you are using this.  So get_implied_value() for example writes to the
cur_slist() and that would cause a warning message and a bug.

I have changed my stuff to use this and since I don't need
__set_fake_cur_slist() any longer, I have removed it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>