comparison: create expr_equal/lte_to_param() functions
Really it's two separate things we want.  If the return value is equal to
a function then we want to know about that.  Or if it's capped by a
parameter then we want to know about that.

Eventually we will probably want to know if the return value is always
greater than a parameter, but that is much less common.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>