comparison: handle unmatched comparisons to " orig" states
The problem here is this:

int returns_less(int x)
{
        int y;

        if (x < 10)
                y = x;
        else
                y = 10;

        return y;
}

On the true branch then we know that "y == x orig".  On the false side,
we have an unmatched state.

For normal variables we take the implied range_list for both sides and
get the comparison from that.  For the " orig" state we can use the work
in smatch_limit.c and call get_orig_estate() using the symbol pointer.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>