user_data: make cross function tracking more specific
The we are processing looks like this:

	copy_from_user(&kernel_struct, p, sizeof(&kernel_struct));

The old way Smatch handled this was that it marked kernel_struct as user
data.  The new way to handle this is to mark each of the members of
kernel_struct as user data.

In the old way, if we wanted to find if kernel_struct.xxx was user data
then we looked at "kernel_struct" but now we can look at
"kernel_struct.xxx" directly.

The second thing that this patch does is it changes how cross function
user_data is tracked.  At the end of the function we want to look at
which elements are user data.  The problem with the old code is that it
didn't distinguish between pointers that were user data because they were
user data when we passed them in and pointers where we copied user data to
them inside the function.  In the new code, we do care about that.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>