flow, clear_buffer: fake assigns were executing unops multiple times
Instead of calling __split_expr() we should just pass the expression
directly to the clients.  That's all we care about.  Passing it to
split_expr() executes the PREOP and POSTOPs more than once and it leads to
false positives about buffer overflows.

There are maybe some related bugs to fix in later patches.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>