db: use comparison in db to determine get_implied_rl().
The key to this patch is that I have changed str_to_rl() to
call_results_to_rl().  That way when we're evaluating

	__smatch_implied_max(strnlen(str, 16));

It can use the comparison to 16 to know the max is 16.

The other changes are just so that we can pass the expression pointer to
call_results_to_rl().

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>