math: fix get_fuzzy_max()
The problem here is if you have something like

	val = 0-u32max (0-u32max, 4-6)

Then the fuzzy min would be 6 and the fuzzy max would be 4 when we call
alloc_rl() it says that that's inverted so it allocates a whole range.
Fuzzy min/max is weird that way.  We need to flip them the right way around
before calling alloc_rl().

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>