comparison: comparisons with 3 variables: "a > b + c"
Smatch comparison is already dealing with the relationships between two
variables so adding a third variable is not that complicated.  You just
need to save a third link so that if any of the variables are modified
then the relationship is destroyed.

The chunk_to_var_sym() is like expr_to_var_sym() except that if there are
two variables involved then it doesn't give you a symbol pointer.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>