db: track some more ==p1 return values
The problem here is that the code thinks it can split apart the return
value into something meaningful because it knows how the function was
called.  But actually the parameter has not been modified within the
function.

So instead of splitting it apart just return that it is equal to a
parameter.

I added some NULL checks for void functions so we don't crash.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>