comparison: partially fix how links are updated
The problem is that a link is a var/sym pair but when you have three
variables then one chunks is not a var/sym pair.  It works fine if you
are creating a new state, in that case you just create three links.  But
where it falls down is when you are inheriting comparisons and don't have
access to the original expression pointer or the var/sym pairs.

Anyway, this fixes it slightly for creating new links.

There is no point in creating a link for a chunk without a sym pointer so
I have removed that.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>