comparison: inherit links when the comparison is set from a function call
If you have:

	a = b;

Then a inherits all the links from b.  If you have:

	if (a < b) {

Then if you know "b < c" then you know that "a < c".  The problem here is
that we are doing:

	a = strnlen(foo, b);

At the end we know that "a <= b" so if "b < c" then "a < c".  It's slightly
different from copying all the links or form the condition.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>