comparison: fix breakage caused by moving to var_sym_lists
I made a mistake in cb978afa810 "comparison: track modifications better
using var_sym_lists" that broke validation/sm_compare11.c.  At the time,
I thought is was an old bug which was only exposed by the new changes.  But
actually it was a bug in the new changes.

We need to store a link from "a + b" => "a + b vs d", even though we don't
have a symbol pointer for "a + b".

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>