check_overflow: silence some "not allocating enough data" false positives
The warning messages from here are mostly caused because you have a call
tree that returns an allocated struct.  The size of the struct changes but
the call tree takes a couple runs of building the database before it is
all updated.

I have changed it to only warn if the function returning the allocated
buffer returns NULL.  This will silence the warnings.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>