kernel data/: update to recent linux-next (3.11-rc1 ish)
It has been a while since I updated the data/ directory.  This is a good
time because things are sort of stable at the end of a merge window.

The one thing which I have not updated is
smatch_data/kernel.expects_err_ptr  which is an empty file for some reason.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>