math: improve how bitwise ANDs are handled
Before it just calculated the the 0-max from both sides and took the
intersection.  In the new way, we take zero, and the min with one bit set
up to the max for both sides and find the intersection.

So before if we had 0xff00 then we took "0-0xff00" but now we do
"0,0x0100-0xff00"  It's still not very accurate really but it's an
improvement.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>