helper: get_member_name() could store unions (sort of)
This is a bit gnarly, but I've had it in my tree for a while so I'm going
to commit it.

The idea is to store union members, in the same way we currently store
struct members.  The problem is that even for named unions Sparse doesn't
store the name so I have to make up a nonsense name.  Right now I have
called it "(union hack last_member)->member".  Hopefully no one sees it
in real life, or if they do the word "hack" lets them know it's not really
the name of the union.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>