db: rename add_returned_state_callback() => add_split_return_callback()
I could never remember the original name and I could never remember what
it did when I saw it in the code.  I guess the reason why it was called
"returned_state" callback is because it is often used to fill the
returned_state table in the database.  But that's not the only use for it.

Really it's like the RETURN_HOOK except that it's split apart and called
multiple times for the same return statement.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>