db, extra, user_data: handle: *x = 1; return x;"
Say you have code like:

int *frob(void)
{
	*x = 42;
	return x;
}

It's good to know to record that *x is 42.  But the real reason for this
is because of memdup_user() which does:

void *memdup_user(void *src, size_t size)
{
	...
	copy_from_user(dest, src, size);
	return dest;
}

In the old code, there was no way to record that the data it was returning
was user data.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>