db: group function pointer return states together
Say you call a function pointer foo->bar(), then maybe there are only two
implementations of ->bar() and they are simple.  In that case, it's useful
to merge all the return states together to see which states are always
true.

The thing is that if the return_ids are the same then they tromp on each
other's data a bit...  I should probably fix that up so first return_id in
a function is very random.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>