data_source: load where data comes from into the database
The scripting bit that produces useful information out of this data isn't
written but I want to cut down on my public/private diff so I'm going to
commit this bit.

This is sort of like param_mapper, I guess.  It maps if you pass one of
your parameters to a different function.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>