buf_size: represent buffer size as a range list instead of a single value
It has been often been a problem where in some cases we want to use the
maximum size of the buffer and in some case we want to use the minimum.  So
internally I've changed it to be a range list and we can ask for the min or
max depending on what we want.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>