buf_size: buffer allocations in the same file are not special
The thought here was that if a buffer is allocated in the file then it's
probably used in the file.  That's probably true, but unless the variable
is static other allocations can be used as well.

In other words this patch means that we say we know the size of the buffer
less often than before.  It should eliminate some false positives.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>