extra: assume indexes are in bounds
Sometimes, we know the size of the array but we don't know the value of the
index.  After this patch we assume that the index is valid.  We only
assume this after we have had a chance to warn about the overflow, so it
shouldn't hide any bugs.  If the index is used more than once though, the
second use doesn't generate a warning.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>