struct_assignment: handle struct assignments to better
When you have structs "foo" and "bar" and the code is:

	foo = bar;

Then create a fake assignment for "foo.x = bar.x" and "foo.y = bar.y".

Eventually I plan to use this code to handle memcpy() better as well.
Currently memcpy() is handled in smatch_clears_buffer() but I will move
it here.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>