user_data: serious bug in handling modifications (missed errors)
The problem was we have re-ordered when modifications show up.  So it
often happened that we would tag something as user data and then handle
that it was modified so now it's not user data any more.

We can't use the modification hooks here for that reason.  I have added
some code in match_assign() to reset the state instead.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>