function_hooks, struct_assignment, extra, db: don't record useless modifications
If we don't know anything about a variable and it is set to something else
which we don't know anything about, then that's not worth recording.  We
might as well not even take any notice that it changed.

The problem here is that the struct_assignment changes made the database
too big.  This change is a way of filtering out useless information.

I worry about this change that it complicates the code but at the same time
I don't see another option because the database was becoming unusable.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>