capped: trace variable capping inside function calls
A lot of code calls to functions like is_valid(x).  If we can put a hard
limit on what valid values are that gets tracked in smatch_param_limit.c
and friends.  But we should also track if the code checks against an
upper bound, even if we don't know what the upper bound is exactly.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>