*new* stored_conditions: make a list of known true/false conditions
This is different from smatch_implied.c because it doesn't know the
values of anything.  It just knows that some condition is true or false
and which variables affect it.  If any of the variables are modified then
the condition is no longer known.

This code will be useful for finding duplicate conditions or maybe for
improving smatch_implied.c.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>