*new* double_checking: warn no duplicative conditons
This test is only enabled with --spammy.  It prints a lot of warnings and
I'm not sure how useful it is to know about duplicate conditions actually.

It doesn't warn about duplicate NULL checks because I've already decided to
ignore people who check for NULL over and over.

One idea would be to make a version this check which only warns about:

	if (foo) {
		if (foo) {

But it wouldn't warn about:

	if (!foo)
		return;
	if (foo) {

I would enable that check without the --spammy flag.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>