avl: allocate on first insert, free on last remove
The ptr_list type allocates on the first insert and free on the last
remove.  Of course, part of the reason ptr_list type does this is because
when you insert then you could insert at the start of the list.  That
isn't true with the AVL list.

But writing it this way will make porting easier so I am doing it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>