stree fallout: implications not working 100%
The problem here is that clone_pool_havers() was not cloning things
properly.  "iter.sm" is a temporary pointer and it's not like
REPLACE_CURRENT_PTR() for slists.

To fix this, I put all the clones into a list and the insert them in a
separate step.

The ->pool variable is necessary for implications to work correctly.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>