extra: fix unknown += assignments
The if we have:

	u64 x;
	u32 y;

	x += y;

Then we don't know the value of x, it could be 0-u64max.  In the original
code we said that it could go up to u32max.  The first part of this
function changes it to take the range from "x" instead of "y".

The next part of this patch says that we should be calculating the
absolute max instead of saying it was unknown anyway.  This mostly makes
the first change irrelevant.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>