flow: fixes for hairy functions
There are some times where Smatch is just not able to parse the code and
has to bail out.  When it does that then it turns off sm_msg().  The
problem is that it doesn't turn it on again after the end of the function.

Also I added a timeout so if the function takes over 5 minutes to parse
then we bail.

I also re-ordered the checks in __split_stmt() but that doesn't change
anything really.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>