stree: fix a bunch of stree leaks
Add a bunch of calls to free_stree().

The change in match_states_stree() from smatch_slist.c is because it's
faster to just append a pointer to the end of a list instead of inserting
it into a stree.  It probably doesn't make a measurable difference, but
it's the thought which counts.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>