slist, returns: introduce merge_stree_no_pools()
The ->pools are used for implications.  In smatch_returns() we want to
merge the all the states together to find the possible return values, but
it doesn't affect code flow or implications.  We don't need to store the
->pools here.

Changing this means we do fewer sm_state allocations and saves memory.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>