missing_break: use fewer states
The check_missing_break.c script uses a lot of states because it records
every variable which is used.  Possibly this is a stupid thing.

We can obviously ignore variables which are not even inside a switch
statement.  This is the change in match_symbol().

The other change is that at the end of a switch statement we can free
all the states.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>