stored_conditions: don't store constants
There are many times where we do:

	do { ... } while (0);

The smatch_stored_conditions.c file is supposed to record conditions so
that we can track them for when they are checked again.  It doesn't make
sense to talk about that for constants.

The other thing is that this was causing some unreachable code warnings.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>