propagate: delete this test
The check_propagate test looked for places in the kernel which did this:

	if (some_func())
		return -EINVAL;

Because it's better kernel style to do:

	ret = some_func();
	if (ret)
		return ret;

So the correct error code is preserved.

But the check had false positives.  Also too many people do this the wrong
way so it generates a lot of warnings.  Also it's a style issue and I would
rather focus on bugs.

Deleted.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>