db: function with too many pointers to it cause slow down
If a function is called too many times then we ignore the callers instead
of trying to merge all the information together.  But the situation here is
that the function had over a thousand pointers to it and it was trying to
merge them all together.

$ smdb fn_ptr blk_add_trace_rq_complete | sed -e 's/\,/\n/g' | wc -l
1813

We have to check the number of pointers before doing get_direct_callers()
so I merged get_function_pointer_callers() into match_data_from_db().

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>