overflow: don't warn if we overflow use buffers
These days Smatch thinks it knows the size of buffers because it records
every allocation.  But some buffers are allocated in user space and Smatch
doesn't know what size they are.  This causes false positives.  Say we have
code like:

	copy_to_user(user_pointer, kernel_pointer, size);

Don't complain that we are writing too much data.  Only complain if we are
reading too much data (info leak).

I probably should write some special code for this case which doesn't use
the database but for now I'm just tired of the false positives so I'm
disabling it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>