avl: fix failing "stree->stree_id == 0" assert
If we are operating on a shallow cloned stree then the stree_id need not
be zero.  This only started showing up when I modified
check_missing_break.c to use fewer states which in turn meant that we
could use more shallow clones.

On my test system I had additional tests enabled so I was creating a lot of
real clones and I didn't see the bug.

The fix is to create a new real clone when the stree_id is not zero.

Reported-by: Silvan Jegen <s.jegen@gmail.com>
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>