unreachable: don't print a warning if we jump to the middle of a loop
There is a some code like this:

	goto first;

	for (;;) {
		XXX_frob();
first:
		more_frob();
	}

Smatch sees the start of the for loop as unreachable.  There isn't really
a good way to fix this so I just ignore any unreachable loops if they are
preceded by a goto.  It's not beautiful, but it's better to not print
false positives.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>