unreachable: add watchdog() to kernel.no_return_funcs.remove
The .remove files in the smatch_data/ directory are all sort of a hack to
manually fix the other data files.  In this case there are two watchdog()
functions in different drives and one returns and one doesn't return.
Treating them both as returning functions doesn't hurt anything but
treating them as unreachable causes a false positive unreachable code
warning.

I had to patch up the script directory to make the .remove file work.

Now all the smatch_scripts/gen_* scripts will take a -p=<project> argument.
Only the gen_no_return_funcs.sh script uses it, the rest ignore it.

no_return is different because it isn't kernel specific and because it's so
important.  That's why it was treated differently before.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>