clear_buffer: remove this code
The smatch_clear_buffer.c was supposed to help with check_rosenberg.c and
also for checking if we used uninitialized variables.  But it sort of
duplicates smatch_param_cleared.c and smatch_struct_assignment.c.  Also the
uninitialized variable code never worked to the point where I could commit
it.  This code was just messing me up.  Delete it for now.

Move the remaining useful bit to smatch_struct_assignment.c.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>